requirements.txt
biopython==1.86
matplotlib==3.10.8
numpy==2.4.3
pandas==3.0.1
torch==2.9.1
tqdm==4.67.1
scikit-learn==1.7.2
